Added skeleton for comp landing page
