+++ src/pages/Competitions/CompetitionLandingPageOld/index.less
@@ -1,6 +1,5 @@
 @import '../../../styles/base.less';
 .CompetitionLandingPage {
-  .noContainer();
   .meta-data {
     font-size: @font2;
     margin-bottom: @margin-base;
@@ -8,10 +7,10 @@
   .meta-data-title {
     margin-bottom: 0px;
   }
-  // .registerbtn {
-  // }
+  .refresh-btn {
+    margin-left: 1rem;
+  }
   p {
     font-size: @font2;
   }
-  @import '../../../styles/impactful.less';
 }
