+++ src/pages/Competitions/CompetitionLeaderboardPage/index.less
@@ -1,6 +1,5 @@
 @import '../../../styles/base.less';
-.CompetitionLandingPage {
-  .noContainer();
+.CompetitionLeaderboardPage {
   .meta-data {
     font-size: @font2;
     margin-bottom: @margin-base;
@@ -8,8 +7,9 @@
   .meta-data-title {
     margin-bottom: 0px;
   }
-  // .registerbtn {
-  // }
+  .refresh-btn {
+    margin-left: 1rem;
+  }
   p {
     font-size: @font2;
   }
